requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
boto3
pymongo
python-dotenv
beautifulsoup4
lxml
python-docx
python-multipart
pdf2docx
pdfplumber
